Update package & module name, add comments, use proper formatting
